+++ example/lib/pages/input_form_field.dart
@@ -51,7 +51,7 @@ class _InputFormFieldPageState extends State<InputFormFieldPage> {
             initialValue: initialValue,
             placeholder: const Text('Name'),
             onChanged: (value) {
-              formKey.currentState!.setValue('username', value);
+              formKey.currentState!.setFieldValue('username', value);
             },
           ),
         ],